Refactor and fix warnings
